requirements.txt
asgiref==3.8.1
Django==5.2.3
django-formtools==2.5.1
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
